requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-socketio==5.3.4
flask-sock==0.7.0
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
propcache==0.3.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
twilio==9.6.5
elevenlabs==1.2.2
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
google-generativeai==0.8.3
google-cloud-speech==2.27.0
openai==1.58.1
